requirements.txt
opencv-python
numpy
mss
pyautogui
pynput
easyocr
torch
stable-baselines3
ultralytics
loguru
pyyaml
